# Add match groups and optimize some