--- a/tremor-value/src/prelude.rs
+++ b/tremor-value/src/prelude.rs
@@ -1,0 +1,18 @@
+// Copyright 2020-2021, The Tremor Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+pub use super::Value;
+pub use simd_json_value::{
+    Builder as BuilderTrait, Mutable as MutableTrait, Vale as ValueTrait, Writable as WritableTrait,
+};