… .vscode/launch.json
… CHANGELOG.md
… bench/match.json.xz
… bench/match.yaml
… bench/match/main.trickle
… bench/real-workflow-throughput-json-v0.10.yaml
… bench/real-workflow-throughput-json-v0.10/main.trickle
… tests/script.rs
… tests/script_error.rs
… tests/script_errors/function_already_defined/error.txt
… tests/script_errors/function_already_defined/script.tremor
… tests/scripts/match_assign/in.xz
… tests/scripts/match_assign/out.xz
… tests/scripts/match_assign/script.tremor
… tests/scripts/match_reorder1/in.xz
… tests/scripts/match_reorder1/out.xz
… tests/scripts/match_reorder1/script.tremor
… tests/scripts/tilde_extractor/in.xz
… tests/scripts/tilde_extractor/out.xz
… tests/scripts/tilde_extractor/script.tremor
… tests/scripts/tilde_extractor_assign/in.xz
… tests/scripts/tilde_extractor_assign/out.xz
… tests/scripts/tilde_extractor_assign/script.tremor
… tremor-cli/src/cli.yaml
… tremor-script/src/ast/eq.rs
… tremor-script/src/grammar.lalrpop
… tremor-script/src/interpreter.rs
… tremor-script/src/interpreter/expr.rs
… tremor-script/src/interpreter/imut_expr.rs
… tremor-script/src/tilde.rs