# Pull request

## Description

This  adds optimization to match statements, they especially target a large number of cases that are quite simple. It also lays the foundation for later optimizations in this space.

## Related

* Related Issues: fixes #834
* [Docs PR](https://github.com/tremor-rs/tremor-www-docs/pull/91)

## Checklist

* [x] The RFC, if required, has been submitted and approved
* [x] Any user-facing impact of the changes is reflected in docs.tremor.rs
* [x] The code is tested
* [x] Use of unsafe code is reasoned about in a comment
* [x] Update CHANGELOG.md appropriately, recording any changes, bug fixes, or other observable changes in behavior
* [x] The performance impact of the change is measured (see below)

## Performance

After discussion, we believe that gaining 200% In complex cases is worth to lose 1 or 2 % of performance on shorter match cases.

### real-workflow-throughput-json
| main | this | delta |
|-|-|-|
|361.6 MB/s|357.3 MB/s| - 1% |

### real-workflow-throughput-json-v0.10
(same logic as `real-workflow-throughput-json` but written in more modern tremor)

| main | this | delta |
|-|-|-|
|368 MB/s|360 MB/s| - 2.2% |

### merge
| main | this | delta |
|-|-|-|
|54.0 MB/s|163.5 MB/s| + 202 % |