@Licenser 2021-03-18
Question: are we OK with a 2-3% perf drop on the baseline test for the perf gain on more complex match statements?