@mfelsche 2021-03-18
the baseline test is a single match arm? or no match at all? Is the performance drop really attributable to your changes? How is the behaviour with middle-sized match statements, say around 10 arms?