@Licenser 2021-03-19
https://github.com/tremor-rs/tremor-runtime/blob/main/bench/real-workflow-throughput-json/main.trickle is the script in question, it is written in a fairly outdated style (we had no pattern matches back then) I'll make a `new` variant to see how that would fare